Clean up minor related `HackAndSlashSweep` action and test
