+++ .Lib9c.Tests/Action/HackAndSlashSweep3Test.cs
@@ -1,5 +1,6 @@
 namespace Lib9c.Tests.Action
 {
+    using System;
     using System.Collections.Generic;
     using System.Linq;
     using Libplanet;
@@ -13,9 +14,9 @@ namespace Lib9c.Tests.Action
     using Nekoyume.Model.State;
     using Nekoyume.TableData;
     using Xunit;
-    using static Lib9c.SerializeKeys;
+    using static SerializeKeys;
 
-    public class HackAndSlashSweepTest1
+    public class HackAndSlashSweep3Test
     {
         private readonly Dictionary<string, string> _sheets;
         private readonly TableSheets _tableSheets;
@@ -35,7 +36,7 @@ namespace Lib9c.Tests.Action
         private readonly IAccountStateDelta _initialState;
         private readonly IRandom _random;
 
-        public HackAndSlashSweepTest1()
+        public HackAndSlashSweep3Test()
         {
             _random = new TestRandom();
             _sheets = TableSheetsImporter.ImportSheets();
@@ -81,6 +82,9 @@ namespace Lib9c.Tests.Action
                     .SetState(Addresses.TableSheet.Derive(key), value.Serialize());
             }
 
+            var arenaSheetAddress = Addresses.GetSheetAddress<ArenaSheet>();
+            _initialState = _initialState.SetState(arenaSheetAddress, null);
+
             foreach (var address in _avatarState.combinationSlotAddresses)
             {
                 var slotState = new CombinationSlotState(
@@ -90,15 +94,34 @@ namespace Lib9c.Tests.Action
             }
         }
 
+        public (List<Guid> Equipments, List<Guid> Costumes) GetDummyItems(AvatarState avatarState)
+        {
+            var equipments = Doomfist.GetAllParts(_tableSheets, avatarState.level)
+                .Select(e => e.NonFungibleId).ToList();
+            var random = new TestRandom();
+            var costumes = new List<Guid>();
+            if (avatarState.level >= GameConfig.RequireCharacterLevel.CharacterFullCostumeSlot)
+            {
+                var costumeId = _tableSheets
+                    .CostumeItemSheet
+                    .Values
+                    .First(r => r.ItemSubType == ItemSubType.FullCostume)
+                    .Id;
+
+                var costume = (Costume)ItemFactory.CreateItem(
+                    _tableSheets.ItemSheet[costumeId], random);
+                avatarState.inventory.AddItem(costume);
+                costumes.Add(costume.ItemId);
+            }
+
+            return (equipments, costumes);
+        }
+
         [Theory]
         [InlineData(1, 1, 1, true)]
         [InlineData(1, 1, 1, false)]
-        [InlineData(2, 1, 30, true)]
-        [InlineData(2, 1, 30, false)]
-        [InlineData(5, 4, 199, false)]
-        [InlineData(5, 4, 199, true)]
-        [InlineData(9, 5, 250, false)]
-        [InlineData(9, 5, 250, true)]
+        [InlineData(2, 1, 2, true)]
+        [InlineData(2, 1, 2, false)]
         public void Execute(int apStoneCount, int worldId, int stageId, bool backward)
         {
             var gameConfigState = _initialState.GetGameConfigState();
@@ -112,6 +135,7 @@ namespace Lib9c.Tests.Action
             {
                 worldInformation =
                     new WorldInformation(0, _initialState.GetSheet<WorldSheet>(), stageId),
+                level = 400,
             };
 
             var row = _tableSheets.MaterialItemSheet.Values.First(r =>
@@ -153,10 +177,14 @@ namespace Lib9c.Tests.Action
 
                 var random = new TestRandom(_random.Seed);
                 var expectedRewardItems =
-                    HackAndSlashSweep1.GetRewardItems(random, playCount, stageRow, _tableSheets.MaterialItemSheet);
+                    HackAndSlashSweep3.GetRewardItems(random, playCount, stageRow, _tableSheets.MaterialItemSheet);
 
-                var action = new HackAndSlashSweep1
+                var (equipments, costumes) = GetDummyItems(avatarState);
+                var action = new HackAndSlashSweep3
                 {
+                    actionPoint = avatarState.actionPoint,
+                    costumes = costumes,
+                    equipments = equipments,
                     avatarAddress = _avatarAddress,
                     apStoneCount = apStoneCount,
                     worldId = worldId,
@@ -174,7 +202,9 @@ namespace Lib9c.Tests.Action
 
                 Assert.Equal(expectedLevel, nextAvatarState.level);
                 Assert.Equal(expectedExp, nextAvatarState.exp);
-                Assert.Equal(expectedRewardItems.Count(), nextAvatarState.inventory.Items.Sum(x => x.count));
+                Assert.Equal(
+                    expectedRewardItems.Count(),
+                    nextAvatarState.inventory.Items.Sum(x => x.count));
                 foreach (var i in nextAvatarState.inventory.Items)
                 {
                     nextAvatarState.inventory.TryGetItem(i.item.Id, out var item);
@@ -188,7 +218,7 @@ namespace Lib9c.Tests.Action
         [InlineData(false)]
         public void Execute_FailedLoadStateException(bool backward)
         {
-            var action = new HackAndSlashSweep1
+            var action = new HackAndSlashSweep3
             {
                 apStoneCount = 1,
                 avatarAddress = _avatarAddress,
@@ -218,7 +248,7 @@ namespace Lib9c.Tests.Action
         [InlineData(100, 1)]
         public void Execute_SheetRowNotFoundException(int worldId, int stageId)
         {
-            var action = new HackAndSlashSweep1
+            var action = new HackAndSlashSweep3
             {
                 apStoneCount = 1,
                 avatarAddress = _avatarAddress,
@@ -238,7 +268,7 @@ namespace Lib9c.Tests.Action
         [InlineData(1, 999)]
         public void Execute_SheetRowColumnException(int worldId, int stageId)
         {
-            var action = new HackAndSlashSweep1
+            var action = new HackAndSlashSweep3
             {
                 apStoneCount = 1,
                 avatarAddress = _avatarAddress,
@@ -255,9 +285,9 @@ namespace Lib9c.Tests.Action
         }
 
         [Fact]
-        public void Execute_InvalidStageException()
+        public void Execute_StageClearedException()
         {
-            var action = new HackAndSlashSweep1
+            var action = new HackAndSlashSweep3
             {
                 apStoneCount = 1,
                 avatarAddress = _avatarAddress,
@@ -265,9 +295,47 @@ namespace Lib9c.Tests.Action
                 stageId = 50,
             };
 
-            Assert.Throws<InvalidStageException>(() => action.Execute(new ActionContext()
+            Assert.Throws<StageNotClearedException>(() => action.Execute(new ActionContext()
             {
                 PreviousStates = _initialState,
+                Signer = _agentAddress,
+                Random = new TestRandom(),
+            }));
+        }
+
+        [Theory]
+        [InlineData(true)]
+        [InlineData(false)]
+        public void Execute_InvalidStageException(bool backward)
+        {
+            var action = new HackAndSlashSweep3
+            {
+                apStoneCount = 1,
+                avatarAddress = _avatarAddress,
+                worldId = 1,
+                stageId = 50,
+            };
+            var worldSheet = _initialState.GetSheet<WorldSheet>();
+            var worldUnlockSheet = _initialState.GetSheet<WorldUnlockSheet>();
+
+            _avatarState.worldInformation.ClearStage(1, 2, 1, worldSheet, worldUnlockSheet);
+
+            var state = _initialState;
+            if (backward)
+            {
+                state = _initialState.SetState(_avatarAddress, _avatarState.Serialize());
+            }
+            else
+            {
+                state = _initialState
+                    .SetState(
+                        _avatarAddress.Derive(LegacyWorldInformationKey),
+                        _avatarState.worldInformation.Serialize());
+            }
+
+            Assert.Throws<InvalidStageException>(() => action.Execute(new ActionContext()
+            {
+                PreviousStates = state,
                 Signer = _agentAddress,
                 Random = new TestRandom(),
             }));
@@ -311,7 +379,7 @@ namespace Lib9c.Tests.Action
                         avatarState.questList.Serialize());
             }
 
-            var action = new HackAndSlashSweep1
+            var action = new HackAndSlashSweep3
             {
                 apStoneCount = 1,
                 avatarAddress = _avatarAddress,
@@ -365,12 +433,12 @@ namespace Lib9c.Tests.Action
                         avatarState.questList.Serialize());
             }
 
-            var action = new HackAndSlashSweep1
+            var action = new HackAndSlashSweep3
             {
                 apStoneCount = apStoneCount,
                 avatarAddress = _avatarAddress,
                 worldId = 1,
-                stageId = 25,
+                stageId = 2,
             };
 
             Assert.Throws<UsageLimitExceedException>(() => action.Execute(new ActionContext()
@@ -397,6 +465,7 @@ namespace Lib9c.Tests.Action
             {
                 worldInformation =
                     new WorldInformation(0, _initialState.GetSheet<WorldSheet>(), 25),
+                level = 400,
             };
 
             var row = _tableSheets.MaterialItemSheet.Values.First(r =>
@@ -426,7 +495,7 @@ namespace Lib9c.Tests.Action
 
             var stageSheet = _initialState.GetSheet<StageSheet>();
             var (expectedLevel, expectedExp) = (0, 0L);
-            if (stageSheet.TryGetValue(25, out var stageRow))
+            if (stageSheet.TryGetValue(2, out var stageRow))
             {
                 var itemPlayCount =
                     gameConfigState.ActionPointMax / stageRow.CostAP * useApStoneCount;
@@ -434,15 +503,20 @@ namespace Lib9c.Tests.Action
                 var playCount = apPlayCount + itemPlayCount;
                 (expectedLevel, expectedExp) = avatarState.GetLevelAndExpV1(
                     _tableSheets.CharacterLevelSheet,
-                    25,
+                    2,
                     playCount);
 
-                var action = new HackAndSlashSweep1
+                var (equipments, costumes) = GetDummyItems(avatarState);
+
+                var action = new HackAndSlashSweep3
                 {
+                    equipments = equipments,
+                    costumes = costumes,
                     avatarAddress = _avatarAddress,
+                    actionPoint = avatarState.actionPoint,
                     apStoneCount = useApStoneCount,
                     worldId = 1,
-                    stageId = 25,
+                    stageId = 2,
                 };
 
                 Assert.Throws<NotEnoughMaterialException>(() => action.Execute(new ActionContext()
@@ -470,6 +544,7 @@ namespace Lib9c.Tests.Action
             {
                 worldInformation =
                     new WorldInformation(0, _initialState.GetSheet<WorldSheet>(), 25),
+                level = 400,
                 actionPoint = 0,
             };
 
@@ -495,7 +570,7 @@ namespace Lib9c.Tests.Action
 
             var stageSheet = _initialState.GetSheet<StageSheet>();
             var (expectedLevel, expectedExp) = (0, 0L);
-            if (stageSheet.TryGetValue(25, out var stageRow))
+            if (stageSheet.TryGetValue(2, out var stageRow))
             {
                 var itemPlayCount =
                     gameConfigState.ActionPointMax / stageRow.CostAP * 1;
@@ -503,24 +578,198 @@ namespace Lib9c.Tests.Action
                 var playCount = apPlayCount + itemPlayCount;
                 (expectedLevel, expectedExp) = avatarState.GetLevelAndExpV1(
                     _tableSheets.CharacterLevelSheet,
-                    25,
+                    2,
                     playCount);
 
-                var action = new HackAndSlashSweep1
+                var (equipments, costumes) = GetDummyItems(avatarState);
+                var action = new HackAndSlashSweep3
                 {
+                    costumes = costumes,
+                    equipments = equipments,
                     avatarAddress = _avatarAddress,
+                    actionPoint = 999999,
                     apStoneCount = 0,
                     worldId = 1,
-                    stageId = 25,
+                    stageId = 2,
                 };
 
-                Assert.Throws<NotEnoughActionPointException>(() => action.Execute(new ActionContext()
-                {
-                    PreviousStates = state,
-                    Signer = _agentAddress,
-                    Random = new TestRandom(),
-                }));
+                Assert.Throws<NotEnoughActionPointException>(() =>
+                    action.Execute(new ActionContext()
+                    {
+                        PreviousStates = state,
+                        Signer = _agentAddress,
+                        Random = new TestRandom(),
+                    }));
             }
+        }
+
+        [Theory]
+        [InlineData(true)]
+        [InlineData(false)]
+        public void Execute_PlayCountIsZeroException(bool backward)
+        {
+            var gameConfigState = _initialState.GetGameConfigState();
+            var avatarState = new AvatarState(
+                _avatarAddress,
+                _agentAddress,
+                0,
+                _initialState.GetAvatarSheets(),
+                gameConfigState,
+                _rankingMapAddress)
+            {
+                worldInformation =
+                    new WorldInformation(0, _initialState.GetSheet<WorldSheet>(), 25),
+                level = 400,
+                actionPoint = 0,
+            };
+
+            IAccountStateDelta state;
+            if (backward)
+            {
+                state = _initialState.SetState(_avatarAddress, avatarState.Serialize());
+            }
+            else
+            {
+                state = _initialState
+                    .SetState(_avatarAddress, avatarState.SerializeV2())
+                    .SetState(
+                        _avatarAddress.Derive(LegacyInventoryKey),
+                        avatarState.inventory.Serialize())
+                    .SetState(
+                        _avatarAddress.Derive(LegacyWorldInformationKey),
+                        avatarState.worldInformation.Serialize())
+                    .SetState(
+                        _avatarAddress.Derive(LegacyQuestListKey),
+                        avatarState.questList.Serialize());
+            }
+
+            var stageSheet = _initialState.GetSheet<StageSheet>();
+            var (expectedLevel, expectedExp) = (0, 0L);
+            if (stageSheet.TryGetValue(2, out var stageRow))
+            {
+                var itemPlayCount =
+                    gameConfigState.ActionPointMax / stageRow.CostAP * 1;
+                var apPlayCount = avatarState.actionPoint / stageRow.CostAP;
+                var playCount = apPlayCount + itemPlayCount;
+                (expectedLevel, expectedExp) = avatarState.GetLevelAndExpV1(
+                    _tableSheets.CharacterLevelSheet,
+                    2,
+                    playCount);
+
+                var (equipments, costumes) = GetDummyItems(avatarState);
+                var action = new HackAndSlashSweep3
+                {
+                    costumes = costumes,
+                    equipments = equipments,
+                    avatarAddress = _avatarAddress,
+                    actionPoint = 0,
+                    apStoneCount = 0,
+                    worldId = 1,
+                    stageId = 2,
+                };
+
+                Assert.Throws<PlayCountIsZeroException>(() =>
+                    action.Execute(new ActionContext()
+                    {
+                        PreviousStates = state,
+                        Signer = _agentAddress,
+                        Random = new TestRandom(),
+                    }));
+            }
+        }
+
+        [Theory]
+        [InlineData(1, 3, true)]
+        [InlineData(1, 3, false)]
+        public void Execute_NotEnoughCombatPointException(int worldId, int stageId, bool backward)
+        {
+            var gameConfigState = _initialState.GetGameConfigState();
+            var avatarState = new AvatarState(
+                _avatarAddress,
+                _agentAddress,
+                0,
+                _initialState.GetAvatarSheets(),
+                gameConfigState,
+                _rankingMapAddress)
+            {
+                worldInformation =
+                    new WorldInformation(0, _initialState.GetSheet<WorldSheet>(), 25),
+                actionPoint = 0,
+                level = 1,
+            };
+
+            IAccountStateDelta state;
+            if (backward)
+            {
+                state = _initialState.SetState(_avatarAddress, avatarState.Serialize());
+            }
+            else
+            {
+                state = _initialState
+                    .SetState(_avatarAddress, avatarState.SerializeV2())
+                    .SetState(
+                        _avatarAddress.Derive(LegacyInventoryKey),
+                        avatarState.inventory.Serialize())
+                    .SetState(
+                        _avatarAddress.Derive(LegacyWorldInformationKey),
+                        avatarState.worldInformation.Serialize())
+                    .SetState(
+                        _avatarAddress.Derive(LegacyQuestListKey),
+                        avatarState.questList.Serialize());
+            }
+
+            var stageSheet = _initialState.GetSheet<StageSheet>();
+            var (expectedLevel, expectedExp) = (0, 0L);
+            if (stageSheet.TryGetValue(stageId, out var stageRow))
+            {
+                var itemPlayCount =
+                    gameConfigState.ActionPointMax / stageRow.CostAP * 1;
+                var apPlayCount = avatarState.actionPoint / stageRow.CostAP;
+                var playCount = apPlayCount + itemPlayCount;
+                (expectedLevel, expectedExp) = avatarState.GetLevelAndExpV1(
+                    _tableSheets.CharacterLevelSheet,
+                    stageId,
+                    playCount);
+
+                var action = new HackAndSlashSweep3
+                {
+                    costumes = new List<Guid>(),
+                    equipments = new List<Guid>(),
+                    avatarAddress = _avatarAddress,
+                    actionPoint = avatarState.actionPoint,
+                    apStoneCount = 1,
+                    worldId = worldId,
+                    stageId = stageId,
+                };
+
+                Assert.Throws<NotEnoughCombatPointException>(() =>
+                    action.Execute(new ActionContext()
+                    {
+                        PreviousStates = state,
+                        Signer = _agentAddress,
+                        Random = new TestRandom(),
+                    }));
+            }
+        }
+
+        [Fact]
+        public void Execute_ActionObsoletedException()
+        {
+            var action = new HackAndSlashSweep3
+            {
+                apStoneCount = 1,
+                avatarAddress = _avatarAddress,
+                worldId = 1,
+                stageId = 50,
+            };
+
+            var state = _initialState.SetState(Addresses.GetSheetAddress<ArenaSheet>(), _tableSheets.ArenaSheet.Serialize());
+            Assert.Throws<ActionObsoletedException>(() => action.Execute(new ActionContext()
+            {
+                PreviousStates = state,
+                Signer = _agentAddress,
+                Random = new TestRandom(),
+            }));
         }
     }
 }
