--- .Lib9c.Tests/Action/HackAndSlashSweepTest4.cs
@@ -1,779 +1,0 @@
-namespace Lib9c.Tests.Action
-{
-    using System;
-    using System.Collections.Generic;
-    using System.Linq;
-    using Libplanet;
-    using Libplanet.Action;
-    using Libplanet.Crypto;
-    using Nekoyume;
-    using Nekoyume.Action;
-    using Nekoyume.Helper;
-    using Nekoyume.Model;
-    using Nekoyume.Model.Item;
-    using Nekoyume.Model.State;
-    using Nekoyume.TableData;
-    using Xunit;
-    using static SerializeKeys;
-
-    public class HackAndSlashSweepTest4
-    {
-        private readonly Dictionary<string, string> _sheets;
-        private readonly TableSheets _tableSheets;
-
-        private readonly Address _agentAddress;
-
-        private readonly Address _avatarAddress;
-        private readonly AvatarState _avatarState;
-
-        private readonly Address _inventoryAddress;
-        private readonly Address _worldInformationAddress;
-        private readonly Address _questListAddress;
-
-        private readonly Address _rankingMapAddress;
-
-        private readonly WeeklyArenaState _weeklyArenaState;
-        private readonly IAccountStateDelta _initialState;
-        private readonly IRandom _random;
-
-        public HackAndSlashSweepTest4()
-        {
-            _random = new TestRandom();
-            _sheets = TableSheetsImporter.ImportSheets();
-            _tableSheets = new TableSheets(_sheets);
-
-            var privateKey = new PrivateKey();
-            _agentAddress = privateKey.PublicKey.ToAddress();
-            var agentState = new AgentState(_agentAddress);
-
-            _avatarAddress = _agentAddress.Derive("avatar");
-            var gameConfigState = new GameConfigState(_sheets[nameof(GameConfigSheet)]);
-            _rankingMapAddress = _avatarAddress.Derive("ranking_map");
-            _avatarState = new AvatarState(
-                _avatarAddress,
-                _agentAddress,
-                0,
-                _tableSheets.GetAvatarSheets(),
-                gameConfigState,
-                _rankingMapAddress
-            )
-            {
-                level = 100,
-            };
-            _inventoryAddress = _avatarAddress.Derive(LegacyInventoryKey);
-            _worldInformationAddress = _avatarAddress.Derive(LegacyWorldInformationKey);
-            _questListAddress = _avatarAddress.Derive(LegacyQuestListKey);
-            agentState.avatarAddresses.Add(0, _avatarAddress);
-
-            _weeklyArenaState = new WeeklyArenaState(0);
-
-            _initialState = new State()
-                .SetState(_weeklyArenaState.address, _weeklyArenaState.Serialize())
-                .SetState(_agentAddress, agentState.SerializeV2())
-                .SetState(_avatarAddress, _avatarState.SerializeV2())
-                .SetState(_inventoryAddress, _avatarState.inventory.Serialize())
-                .SetState(_worldInformationAddress, _avatarState.worldInformation.Serialize())
-                .SetState(_questListAddress, _avatarState.questList.Serialize())
-                .SetState(gameConfigState.address, gameConfigState.Serialize());
-
-            foreach (var (key, value) in _sheets)
-            {
-                _initialState = _initialState
-                    .SetState(Addresses.TableSheet.Derive(key), value.Serialize());
-            }
-
-            var arenaSheetAddress = Addresses.GetSheetAddress<ArenaSheet>();
-            _initialState = _initialState.SetState(arenaSheetAddress, null);
-
-            foreach (var address in _avatarState.combinationSlotAddresses)
-            {
-                var slotState = new CombinationSlotState(
-                    address,
-                    GameConfig.RequireClearedStageLevel.CombinationEquipmentAction);
-                _initialState = _initialState.SetState(address, slotState.Serialize());
-            }
-        }
-
-        public (List<Guid> Equipments, List<Guid> Costumes) GetDummyItems(AvatarState avatarState)
-        {
-            var equipments = Doomfist.GetAllParts(_tableSheets, avatarState.level)
-                .Select(e => e.NonFungibleId).ToList();
-            var random = new TestRandom();
-            var costumes = new List<Guid>();
-            if (avatarState.level >= GameConfig.RequireCharacterLevel.CharacterFullCostumeSlot)
-            {
-                var costumeId = _tableSheets
-                    .CostumeItemSheet
-                    .Values
-                    .First(r => r.ItemSubType == ItemSubType.FullCostume)
-                    .Id;
-
-                var costume = (Costume)ItemFactory.CreateItem(
-                    _tableSheets.ItemSheet[costumeId], random);
-                avatarState.inventory.AddItem(costume);
-                costumes.Add(costume.ItemId);
-            }
-
-            return (equipments, costumes);
-        }
-
-        [Theory]
-        [InlineData(1, 1, 1, false, true)]
-        [InlineData(1, 1, 1, false, false)]
-        [InlineData(2, 1, 2, false, true)]
-        [InlineData(2, 1, 2, false, false)]
-        [InlineData(2, 2, 51, false, true)]
-        [InlineData(2, 2, 51, false, false)]
-        [InlineData(2, 2, 52, false, true)]
-        [InlineData(2, 2, 52, false, false)]
-        [InlineData(2, 1, 1, true, true)]
-        [InlineData(2, 1, 1, true, false)]
-        [InlineData(2, 1, 2, true, true)]
-        [InlineData(2, 1, 2, true, false)]
-        [InlineData(2, 2, 51, true, true)]
-        [InlineData(2, 2, 51, true, false)]
-        [InlineData(2, 2, 52, true, true)]
-        [InlineData(2, 2, 52, true, false)]
-        public void Execute(int apStoneCount, int worldId, int stageId, bool challenge, bool backward)
-        {
-            var gameConfigState = _initialState.GetGameConfigState();
-            var prevStageId = stageId - 1;
-            var worldInformation = new WorldInformation(
-                    0, _initialState.GetSheet<WorldSheet>(), challenge ? prevStageId : stageId);
-
-            if (challenge)
-            {
-                worldInformation.UnlockWorld(worldId, 0,  _tableSheets.WorldSheet);
-            }
-
-            var avatarState = new AvatarState(
-                _avatarAddress,
-                _agentAddress,
-                0,
-                _initialState.GetAvatarSheets(),
-                gameConfigState,
-                _rankingMapAddress)
-            {
-                worldInformation = worldInformation,
-                level = 400,
-            };
-
-            var row = _tableSheets.MaterialItemSheet.Values.First(r =>
-                r.ItemSubType == ItemSubType.ApStone);
-            var apStone = ItemFactory.CreateTradableMaterial(row);
-            avatarState.inventory.AddItem(apStone, apStoneCount);
-
-            IAccountStateDelta state;
-            if (backward)
-            {
-                state = _initialState.SetState(_avatarAddress, avatarState.Serialize());
-            }
-            else
-            {
-                state = _initialState
-                    .SetState(_avatarAddress, avatarState.SerializeV2())
-                    .SetState(
-                        _avatarAddress.Derive(LegacyInventoryKey),
-                        avatarState.inventory.Serialize())
-                    .SetState(
-                        _avatarAddress.Derive(LegacyWorldInformationKey),
-                        avatarState.worldInformation.Serialize())
-                    .SetState(
-                        _avatarAddress.Derive(LegacyQuestListKey),
-                        avatarState.questList.Serialize());
-            }
-
-            var stageSheet = _initialState.GetSheet<StageSheet>();
-            var (expectedLevel, expectedExp) = (0, 0L);
-            if (stageSheet.TryGetValue(stageId, out var stageRow))
-            {
-                var itemPlayCount = gameConfigState.ActionPointMax / stageRow.CostAP * apStoneCount;
-                var apPlayCount = avatarState.actionPoint / stageRow.CostAP;
-                var playCount = apPlayCount + itemPlayCount;
-                (expectedLevel, expectedExp) = avatarState.GetLevelAndExp(
-                    _tableSheets.CharacterLevelSheet,
-                    stageId,
-                    playCount);
-
-                var random = new TestRandom(_random.Seed);
-                var expectedRewardItems =
-                    HackAndSlashSweep4.GetRewardItems(random, playCount, stageRow, _tableSheets.MaterialItemSheet);
-
-                var (equipments, costumes) = GetDummyItems(avatarState);
-                var action = new HackAndSlashSweep4
-                {
-                    actionPoint = avatarState.actionPoint,
-                    costumes = costumes,
-                    equipments = equipments,
-                    avatarAddress = _avatarAddress,
-                    apStoneCount = apStoneCount,
-                    worldId = worldId,
-                    stageId = stageId,
-                };
-
-                state = action.Execute(new ActionContext
-                {
-                    PreviousStates = state,
-                    Signer = _agentAddress,
-                    Random = _random,
-                });
-
-                var nextAvatarState = state.GetAvatarStateV2(_avatarAddress);
-
-                Assert.Equal(expectedLevel, nextAvatarState.level);
-                Assert.Equal(expectedExp, nextAvatarState.exp);
-                Assert.Equal(
-                    expectedRewardItems.Count(),
-                    nextAvatarState.inventory.Items.Sum(x => x.count));
-                foreach (var i in nextAvatarState.inventory.Items)
-                {
-                    nextAvatarState.inventory.TryGetItem(i.item.Id, out var item);
-                    Assert.Equal(expectedRewardItems.Count(x => x.Id == i.item.Id), item.count);
-                }
-            }
-        }
-
-        [Theory]
-        [InlineData(true)]
-        [InlineData(false)]
-        public void Execute_FailedLoadStateException(bool backward)
-        {
-            var action = new HackAndSlashSweep4
-            {
-                apStoneCount = 1,
-                avatarAddress = _avatarAddress,
-                worldId = 1,
-                stageId = 1,
-            };
-
-            var state = backward ? new State() : _initialState;
-            if (!backward)
-            {
-                state = _initialState
-                    .SetState(_avatarAddress, _avatarState.SerializeV2())
-                    .SetState(_avatarAddress.Derive(LegacyInventoryKey), null!)
-                    .SetState(_avatarAddress.Derive(LegacyWorldInformationKey), null!)
-                    .SetState(_avatarAddress.Derive(LegacyQuestListKey), null!);
-            }
-
-            Assert.Throws<FailedLoadStateException>(() => action.Execute(new ActionContext()
-            {
-                PreviousStates = state,
-                Signer = _agentAddress,
-                Random = new TestRandom(),
-            }));
-        }
-
-        [Theory]
-        [InlineData(100, 1)]
-        public void Execute_SheetRowNotFoundException(int worldId, int stageId)
-        {
-            var action = new HackAndSlashSweep4
-            {
-                apStoneCount = 1,
-                avatarAddress = _avatarAddress,
-                worldId = worldId,
-                stageId = stageId,
-            };
-
-            Assert.Throws<SheetRowNotFoundException>(() => action.Execute(new ActionContext()
-            {
-                PreviousStates = _initialState,
-                Signer = _agentAddress,
-                Random = new TestRandom(),
-            }));
-        }
-
-        [Theory]
-        [InlineData(1, 999)]
-        [InlineData(2, 50)]
-        public void Execute_SheetRowColumnException(int worldId, int stageId)
-        {
-            var action = new HackAndSlashSweep4
-            {
-                apStoneCount = 1,
-                avatarAddress = _avatarAddress,
-                worldId = worldId,
-                stageId = stageId,
-            };
-
-            Assert.Throws<SheetRowColumnException>(() => action.Execute(new ActionContext()
-            {
-                PreviousStates = _initialState,
-                Signer = _agentAddress,
-                Random = new TestRandom(),
-            }));
-        }
-
-        [Theory]
-        [InlineData(1, 48, 1, 50, true)]
-        [InlineData(1, 48, 1, 50, false)]
-        [InlineData(1, 49, 2, 51, true)]
-        [InlineData(1, 49, 2, 51, false)]
-        public void Execute_InvalidStageException(int clearedWorldId, int clearedStageId, int worldId, int stageId, bool backward)
-        {
-            var action = new HackAndSlashSweep4
-            {
-                apStoneCount = 1,
-                avatarAddress = _avatarAddress,
-                worldId = worldId,
-                stageId = stageId,
-            };
-            var worldSheet = _initialState.GetSheet<WorldSheet>();
-            var worldUnlockSheet = _initialState.GetSheet<WorldUnlockSheet>();
-
-            _avatarState.worldInformation.ClearStage(clearedWorldId, clearedStageId, 1, worldSheet, worldUnlockSheet);
-
-            var state = _initialState;
-            if (backward)
-            {
-                state = _initialState.SetState(_avatarAddress, _avatarState.Serialize());
-            }
-            else
-            {
-                state = _initialState
-                    .SetState(
-                        _avatarAddress.Derive(LegacyWorldInformationKey),
-                        _avatarState.worldInformation.Serialize());
-            }
-
-            Assert.Throws<InvalidStageException>(() => action.Execute(new ActionContext()
-            {
-                PreviousStates = state,
-                Signer = _agentAddress,
-                Random = new TestRandom(),
-            }));
-        }
-
-        [Theory]
-        [InlineData(GameConfig.MimisbrunnrWorldId, true)]
-        [InlineData(GameConfig.MimisbrunnrWorldId, false)]
-        public void Execute_InvalidWorldException(int worldId, bool backward)
-        {
-            var gameConfigState = new GameConfigState(_sheets[nameof(GameConfigSheet)]);
-            var avatarState = new AvatarState(
-                _avatarAddress,
-                _agentAddress,
-                0,
-                _initialState.GetAvatarSheets(),
-                gameConfigState,
-                _rankingMapAddress)
-            {
-                worldInformation =
-                    new WorldInformation(0, _initialState.GetSheet<WorldSheet>(), 10000001),
-            };
-
-            IAccountStateDelta state;
-            if (backward)
-            {
-                state = _initialState.SetState(_avatarAddress, avatarState.Serialize());
-            }
-            else
-            {
-                state = _initialState
-                    .SetState(_avatarAddress, avatarState.SerializeV2())
-                    .SetState(
-                        _avatarAddress.Derive(LegacyInventoryKey),
-                        avatarState.inventory.Serialize())
-                    .SetState(
-                        _avatarAddress.Derive(LegacyWorldInformationKey),
-                        avatarState.worldInformation.Serialize())
-                    .SetState(
-                        _avatarAddress.Derive(LegacyQuestListKey),
-                        avatarState.questList.Serialize());
-            }
-
-            var action = new HackAndSlashSweep4
-            {
-                apStoneCount = 1,
-                avatarAddress = _avatarAddress,
-                worldId = worldId,
-                stageId = 10000001,
-            };
-
-            Assert.Throws<InvalidWorldException>(() => action.Execute(new ActionContext()
-            {
-                PreviousStates = state,
-                Signer = _agentAddress,
-                Random = new TestRandom(),
-            }));
-        }
-
-        [Theory]
-        [InlineData(99, true)]
-        [InlineData(99, false)]
-        public void Execute_UsageLimitExceedException(int apStoneCount, bool backward)
-        {
-            var gameConfigState = new GameConfigState(_sheets[nameof(GameConfigSheet)]);
-            var avatarState = new AvatarState(
-                _avatarAddress,
-                _agentAddress,
-                0,
-                _initialState.GetAvatarSheets(),
-                gameConfigState,
-                _rankingMapAddress)
-            {
-                worldInformation =
-                    new WorldInformation(0, _initialState.GetSheet<WorldSheet>(), 25),
-            };
-
-            IAccountStateDelta state;
-            if (backward)
-            {
-                state = _initialState.SetState(_avatarAddress, avatarState.Serialize());
-            }
-            else
-            {
-                state = _initialState
-                    .SetState(_avatarAddress, avatarState.SerializeV2())
-                    .SetState(
-                        _avatarAddress.Derive(LegacyInventoryKey),
-                        avatarState.inventory.Serialize())
-                    .SetState(
-                        _avatarAddress.Derive(LegacyWorldInformationKey),
-                        avatarState.worldInformation.Serialize())
-                    .SetState(
-                        _avatarAddress.Derive(LegacyQuestListKey),
-                        avatarState.questList.Serialize());
-            }
-
-            var action = new HackAndSlashSweep4
-            {
-                apStoneCount = apStoneCount,
-                avatarAddress = _avatarAddress,
-                worldId = 1,
-                stageId = 2,
-            };
-
-            Assert.Throws<UsageLimitExceedException>(() => action.Execute(new ActionContext()
-            {
-                PreviousStates = state,
-                Signer = _agentAddress,
-                Random = new TestRandom(),
-            }));
-        }
-
-        [Theory]
-        [InlineData(3, 2, true)]
-        [InlineData(7, 5, false)]
-        public void Execute_NotEnoughMaterialException(int useApStoneCount, int holdingApStoneCount, bool backward)
-        {
-            var gameConfigState = _initialState.GetGameConfigState();
-            var avatarState = new AvatarState(
-                _avatarAddress,
-                _agentAddress,
-                0,
-                _initialState.GetAvatarSheets(),
-                gameConfigState,
-                _rankingMapAddress)
-            {
-                worldInformation =
-                    new WorldInformation(0, _initialState.GetSheet<WorldSheet>(), 25),
-                level = 400,
-            };
-
-            var row = _tableSheets.MaterialItemSheet.Values.First(r =>
-                r.ItemSubType == ItemSubType.ApStone);
-            var apStone = ItemFactory.CreateTradableMaterial(row);
-            avatarState.inventory.AddItem(apStone, holdingApStoneCount);
-
-            IAccountStateDelta state;
-            if (backward)
-            {
-                state = _initialState.SetState(_avatarAddress, avatarState.Serialize());
-            }
-            else
-            {
-                state = _initialState
-                    .SetState(_avatarAddress, avatarState.SerializeV2())
-                    .SetState(
-                        _avatarAddress.Derive(LegacyInventoryKey),
-                        avatarState.inventory.Serialize())
-                    .SetState(
-                        _avatarAddress.Derive(LegacyWorldInformationKey),
-                        avatarState.worldInformation.Serialize())
-                    .SetState(
-                        _avatarAddress.Derive(LegacyQuestListKey),
-                        avatarState.questList.Serialize());
-            }
-
-            var stageSheet = _initialState.GetSheet<StageSheet>();
-            var (expectedLevel, expectedExp) = (0, 0L);
-            if (stageSheet.TryGetValue(2, out var stageRow))
-            {
-                var itemPlayCount =
-                    gameConfigState.ActionPointMax / stageRow.CostAP * useApStoneCount;
-                var apPlayCount = avatarState.actionPoint / stageRow.CostAP;
-                var playCount = apPlayCount + itemPlayCount;
-                (expectedLevel, expectedExp) = avatarState.GetLevelAndExp(
-                    _tableSheets.CharacterLevelSheet,
-                    2,
-                    playCount);
-
-                var (equipments, costumes) = GetDummyItems(avatarState);
-
-                var action = new HackAndSlashSweep4
-                {
-                    equipments = equipments,
-                    costumes = costumes,
-                    avatarAddress = _avatarAddress,
-                    actionPoint = avatarState.actionPoint,
-                    apStoneCount = useApStoneCount,
-                    worldId = 1,
-                    stageId = 2,
-                };
-
-                Assert.Throws<NotEnoughMaterialException>(() => action.Execute(new ActionContext()
-                {
-                    PreviousStates = state,
-                    Signer = _agentAddress,
-                    Random = new TestRandom(),
-                }));
-            }
-        }
-
-        [Theory]
-        [InlineData(true)]
-        [InlineData(false)]
-        public void Execute_NotEnoughActionPointException(bool backward)
-        {
-            var gameConfigState = _initialState.GetGameConfigState();
-            var avatarState = new AvatarState(
-                _avatarAddress,
-                _agentAddress,
-                0,
-                _initialState.GetAvatarSheets(),
-                gameConfigState,
-                _rankingMapAddress)
-            {
-                worldInformation =
-                    new WorldInformation(0, _initialState.GetSheet<WorldSheet>(), 25),
-                level = 400,
-                actionPoint = 0,
-            };
-
-            IAccountStateDelta state;
-            if (backward)
-            {
-                state = _initialState.SetState(_avatarAddress, avatarState.Serialize());
-            }
-            else
-            {
-                state = _initialState
-                    .SetState(_avatarAddress, avatarState.SerializeV2())
-                    .SetState(
-                        _avatarAddress.Derive(LegacyInventoryKey),
-                        avatarState.inventory.Serialize())
-                    .SetState(
-                        _avatarAddress.Derive(LegacyWorldInformationKey),
-                        avatarState.worldInformation.Serialize())
-                    .SetState(
-                        _avatarAddress.Derive(LegacyQuestListKey),
-                        avatarState.questList.Serialize());
-            }
-
-            var stageSheet = _initialState.GetSheet<StageSheet>();
-            var (expectedLevel, expectedExp) = (0, 0L);
-            if (stageSheet.TryGetValue(2, out var stageRow))
-            {
-                var itemPlayCount =
-                    gameConfigState.ActionPointMax / stageRow.CostAP * 1;
-                var apPlayCount = avatarState.actionPoint / stageRow.CostAP;
-                var playCount = apPlayCount + itemPlayCount;
-                (expectedLevel, expectedExp) = avatarState.GetLevelAndExp(
-                    _tableSheets.CharacterLevelSheet,
-                    2,
-                    playCount);
-
-                var (equipments, costumes) = GetDummyItems(avatarState);
-                var action = new HackAndSlashSweep4
-                {
-                    costumes = costumes,
-                    equipments = equipments,
-                    avatarAddress = _avatarAddress,
-                    actionPoint = 999999,
-                    apStoneCount = 0,
-                    worldId = 1,
-                    stageId = 2,
-                };
-
-                Assert.Throws<NotEnoughActionPointException>(() =>
-                    action.Execute(new ActionContext()
-                    {
-                        PreviousStates = state,
-                        Signer = _agentAddress,
-                        Random = new TestRandom(),
-                    }));
-            }
-        }
-
-        [Theory]
-        [InlineData(true)]
-        [InlineData(false)]
-        public void Execute_PlayCountIsZeroException(bool backward)
-        {
-            var gameConfigState = _initialState.GetGameConfigState();
-            var avatarState = new AvatarState(
-                _avatarAddress,
-                _agentAddress,
-                0,
-                _initialState.GetAvatarSheets(),
-                gameConfigState,
-                _rankingMapAddress)
-            {
-                worldInformation =
-                    new WorldInformation(0, _initialState.GetSheet<WorldSheet>(), 25),
-                level = 400,
-                actionPoint = 0,
-            };
-
-            IAccountStateDelta state;
-            if (backward)
-            {
-                state = _initialState.SetState(_avatarAddress, avatarState.Serialize());
-            }
-            else
-            {
-                state = _initialState
-                    .SetState(_avatarAddress, avatarState.SerializeV2())
-                    .SetState(
-                        _avatarAddress.Derive(LegacyInventoryKey),
-                        avatarState.inventory.Serialize())
-                    .SetState(
-                        _avatarAddress.Derive(LegacyWorldInformationKey),
-                        avatarState.worldInformation.Serialize())
-                    .SetState(
-                        _avatarAddress.Derive(LegacyQuestListKey),
-                        avatarState.questList.Serialize());
-            }
-
-            var stageSheet = _initialState.GetSheet<StageSheet>();
-            var (expectedLevel, expectedExp) = (0, 0L);
-            if (stageSheet.TryGetValue(2, out var stageRow))
-            {
-                var itemPlayCount =
-                    gameConfigState.ActionPointMax / stageRow.CostAP * 1;
-                var apPlayCount = avatarState.actionPoint / stageRow.CostAP;
-                var playCount = apPlayCount + itemPlayCount;
-                (expectedLevel, expectedExp) = avatarState.GetLevelAndExp(
-                    _tableSheets.CharacterLevelSheet,
-                    2,
-                    playCount);
-
-                var (equipments, costumes) = GetDummyItems(avatarState);
-                var action = new HackAndSlashSweep4
-                {
-                    costumes = costumes,
-                    equipments = equipments,
-                    avatarAddress = _avatarAddress,
-                    actionPoint = 0,
-                    apStoneCount = 0,
-                    worldId = 1,
-                    stageId = 2,
-                };
-
-                Assert.Throws<PlayCountIsZeroException>(() =>
-                    action.Execute(new ActionContext()
-                    {
-                        PreviousStates = state,
-                        Signer = _agentAddress,
-                        Random = new TestRandom(),
-                    }));
-            }
-        }
-
-        [Theory]
-        [InlineData(1, 3, true)]
-        [InlineData(1, 3, false)]
-        public void Execute_NotEnoughCombatPointException(int worldId, int stageId, bool backward)
-        {
-            var gameConfigState = _initialState.GetGameConfigState();
-            var avatarState = new AvatarState(
-                _avatarAddress,
-                _agentAddress,
-                0,
-                _initialState.GetAvatarSheets(),
-                gameConfigState,
-                _rankingMapAddress)
-            {
-                worldInformation =
-                    new WorldInformation(0, _initialState.GetSheet<WorldSheet>(), 25),
-                actionPoint = 0,
-                level = 1,
-            };
-
-            IAccountStateDelta state;
-            if (backward)
-            {
-                state = _initialState.SetState(_avatarAddress, avatarState.Serialize());
-            }
-            else
-            {
-                state = _initialState
-                    .SetState(_avatarAddress, avatarState.SerializeV2())
-                    .SetState(
-                        _avatarAddress.Derive(LegacyInventoryKey),
-                        avatarState.inventory.Serialize())
-                    .SetState(
-                        _avatarAddress.Derive(LegacyWorldInformationKey),
-                        avatarState.worldInformation.Serialize())
-                    .SetState(
-                        _avatarAddress.Derive(LegacyQuestListKey),
-                        avatarState.questList.Serialize());
-            }
-
-            var stageSheet = _initialState.GetSheet<StageSheet>();
-            var (expectedLevel, expectedExp) = (0, 0L);
-            if (stageSheet.TryGetValue(stageId, out var stageRow))
-            {
-                var itemPlayCount =
-                    gameConfigState.ActionPointMax / stageRow.CostAP * 1;
-                var apPlayCount = avatarState.actionPoint / stageRow.CostAP;
-                var playCount = apPlayCount + itemPlayCount;
-                (expectedLevel, expectedExp) = avatarState.GetLevelAndExp(
-                    _tableSheets.CharacterLevelSheet,
-                    stageId,
-                    playCount);
-
-                var action = new HackAndSlashSweep4
-                {
-                    costumes = new List<Guid>(),
-                    equipments = new List<Guid>(),
-                    avatarAddress = _avatarAddress,
-                    actionPoint = avatarState.actionPoint,
-                    apStoneCount = 1,
-                    worldId = worldId,
-                    stageId = stageId,
-                };
-
-                Assert.Throws<NotEnoughCombatPointException>(() =>
-                    action.Execute(new ActionContext()
-                    {
-                        PreviousStates = state,
-                        Signer = _agentAddress,
-                        Random = new TestRandom(),
-                    }));
-            }
-        }
-
-        [Fact]
-        public void Execute_ActionObsoletedException()
-        {
-            var action = new HackAndSlashSweep4
-            {
-                apStoneCount = 1,
-                avatarAddress = _avatarAddress,
-                worldId = 1,
-                stageId = 50,
-            };
-
-            var state = _initialState.SetState(Addresses.GetSheetAddress<ArenaSheet>(), _tableSheets.ArenaSheet.Serialize());
-            Assert.Throws<ActionObsoletedException>(() => action.Execute(new ActionContext()
-            {
-                PreviousStates = state,
-                Signer = _agentAddress,
-                Random = new TestRandom(),
-            }));
-        }
-    }
-}
