+++ Lib9c/Action/HackAndSlashSweep4.cs
@@ -15,11 +15,11 @@ using static Lib9c.SerializeKeys;
 
 namespace Nekoyume.Action
 {
-    [Serializable]
     /// <summary>
     /// Introduced at https://github.com/planetarium/lib9c/pull/1017
     /// Updated at https://github.com/planetarium/lib9c/pull/1176
     /// </summary>
+    [Serializable]
     [ActionType("hack_and_slash_sweep4")]
     public class HackAndSlashSweep4 : GameAction
     {
